Redesigned student dashboard
