+++ qresent/lib/screens/admin_dashboard.dart
@@ -30,6 +30,7 @@ class _AdminDashboardState extends State<AdminDashboard> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
+        centerTitle: true,
         title: const Text(
           "Dashboard",
         ),
