+++ qresent/lib/screens/student_home.dart
@@ -2,6 +2,8 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
+import 'package:qresent/model/course_model.dart';
+import 'package:qresent/model/user_model.dart';
 import 'package:qresent/screens/login_screen.dart';
 import 'scan_qr.dart';
 
@@ -14,16 +16,34 @@ class HomePage extends StatefulWidget {
 
 class _HomePage extends State<HomePage> {
   final CollectionReference coursesRef =
+      FirebaseFirestore.instance.collection("Courses");
+  final CollectionReference teachersRef =
       FirebaseFirestore.instance.collection("Users");
+  final CollectionReference attendancesRef =
+      FirebaseFirestore.instance.collection("Attendance");
+
+  final TextEditingController _searchController = TextEditingController();
+
+  User? user = FirebaseAuth.instance.currentUser;
+  List<CourseModel> _coursesList = [];
+  List<CourseModel> _resultsList = [];
+  Map<CourseModel, List<String>> intervals = {};
   final FirebaseAuth _auth = FirebaseAuth.instance;
-  List<dynamic> _courses = [];
   List<String> courses = [];
   var stringWord = '';
 
   @override
   void initState() {
     getCourses();
+    _searchController.addListener(_onSearchChanged);
     super.initState();
+  }
+
+  @override
+  void dispose() {
+    _searchController.removeListener(_onSearchChanged);
+    _searchController.dispose();
+    super.dispose();
   }
 
   Future<void> signOut() async {
@@ -37,81 +57,209 @@ class _HomePage extends State<HomePage> {
   }
 
   getCourses() async {
-    final User? user = _auth.currentUser;
-    final uid = user?.uid;
+    List<CourseModel> courseListTemp = [];
+    UserModel teacher = UserModel();
+    await teachersRef
+        .where("Email", isEqualTo: user?.email)
+        .get()
+        .then((QuerySnapshot snapshot) {
+      teacher = UserModel.fromMap(snapshot.docs[0].data());
+    });
 
-    DocumentSnapshot userData =
-        await FirebaseFirestore.instance.collection("Users").doc(uid).get();
+    await coursesRef
+        .where("UID", whereIn: teacher.assignedCourses)
+        .get()
+        .then((QuerySnapshot snapshot) {
+      for (var documentSnapshot in snapshot.docs) {
+        courseListTemp.add(CourseModel.fromMap(documentSnapshot.data()));
+      }
+    });
 
     setState(() {
-      _courses = userData["AssignedCourses"];
-      courses = _courses.cast<String>();
-      stringWord = courses.join("\n");
+      _coursesList = courseListTemp;
     });
+    for (CourseModel course in _coursesList) {
+      await getIntervals(course);
+    }
+    searchResultList();
+  }
+
+  searchResultList() {
+    List<CourseModel> showResults = [];
+
+    if (_searchController.text != "") {
+      for (var doc in _coursesList) {
+        var courseName = doc.uid.toLowerCase();
+
+        if (courseName.contains(_searchController.text.toLowerCase())) {
+          showResults.add(doc);
+        }
+      }
+    } else {
+      showResults = List.from(_coursesList);
+    }
+
+    setState(() {
+      _resultsList = showResults;
+    });
+  }
+
+  _onSearchChanged() {
+    searchResultList();
+  }
+
+  getIntervals(CourseModel course) async {
+    List<String> tempIntervals = [];
+    await coursesRef
+        .doc(course.uid)
+        .get()
+        .then((value) => tempIntervals = List.from(value.get("Intervals")));
+    setState(() {
+      intervals[course] = tempIntervals;
+    });
+  }
+
+  createInfoAlertDialog(BuildContext context, CourseModel course) {
+    return showDialog(
+      context: context,
+      builder: (context) {
+        return AlertDialog(
+          title: Text(
+            course.uid,
+            textAlign: TextAlign.center,
+          ),
+          content: Text(course.information!),
+          actions: [
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                TextButton(
+                  onPressed: () {
+                    Navigator.of(context).pop();
+                  },
+                  child: const Text("Exit"),
+                ),
+              ],
+            )
+          ],
+        );
+      },
+    );
   }
 
   @override
   Widget build(BuildContext context) {
-    final ButtonStyle flatButtonStyle = TextButton.styleFrom(
-      primary: Colors.black,
-      minimumSize: const Size(88, 44),
-      padding: const EdgeInsets.symmetric(horizontal: 16.0),
-      shape: const RoundedRectangleBorder(
-        borderRadius: BorderRadius.all(Radius.circular(2.0)),
-      ),
-      backgroundColor: Colors.blue,
-    );
     return Scaffold(
-        appBar: AppBar(
-          centerTitle: true,
-          title: const Text("Home"),
-          actions: [
-            IconButton(
-              onPressed: () {
-                signOut();
-              },
-              icon: const Icon(Icons.logout),
-            )
-          ],
-        ),
-        body: Center(
+      appBar: AppBar(
+        centerTitle: true,
+        title: const Text("Home"),
+        actions: [
+          IconButton(
+            onPressed: () {
+              signOut();
+            },
+            icon: const Icon(Icons.logout),
+          )
+        ],
+      ),
+      floatingActionButton: FloatingActionButton.extended(
+        label: const Text("Scan QR"),
+        icon: const Icon(Icons.camera_alt),
+        onPressed: () {
+          _showScan();
+        },
+      ),
+      floatingActionButtonLocation: FloatingActionButtonLocation.centerFloat,
+      body: Center(
+        child: Container(
+          width: 500,
+          padding: const EdgeInsets.only(
+            left: 10,
+            right: 10,
+            top: 10,
+          ),
           child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
             children: <Widget>[
-              TextButton(
-                style: flatButtonStyle,
-                child: const Text("Courses"),
-                onPressed: () {
-                  _showCourses();
-                },
+              Container(
+                padding: const EdgeInsets.fromLTRB(5, 0, 5, 0),
+                child: TextField(
+                  controller: _searchController,
+                  decoration: const InputDecoration(
+                    hintText: 'Search a course',
+                    border: OutlineInputBorder(
+                      borderRadius: BorderRadius.all(Radius.circular(35)),
+                    ),
+                    prefixIcon: Icon(Icons.search),
+                  ),
+                ),
               ),
-              TextButton(
-                  style: flatButtonStyle,
-                  onPressed: () {
-                    _showScan();
+              Expanded(
+                child: ListView.builder(
+                  itemCount: _resultsList.length,
+                  itemBuilder: (context, index) {
+                    return Card(
+                      elevation: 4,
+                      child: ListTile(
+                        title: Row(
+                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                          children: [
+                            Row(
+                              children: [
+                                Text(_resultsList[index].uid),
+                              ],
+                            ),
+                            Row(
+                              children: [
+                                IconButton(
+                                  onPressed: () {
+                                    createInfoAlertDialog(
+                                        context, _resultsList[index]);
+                                  },
+                                  alignment: Alignment.center,
+                                  icon: const Icon(
+                                    Icons.info,
+                                    color: Colors.blueAccent,
+                                    size: 32,
+                                  ),
+                                ),
+                              ],
+                            ),
+                          ],
+                        ),
+                        subtitle: ListView.builder(
+                          physics: const ClampingScrollPhysics(),
+                          shrinkWrap: true,
+                          itemCount: intervals[_resultsList[index]]!.length,
+                          itemBuilder: (context, intervalIndex) {
+                            return Card(
+                              elevation: 4,
+                              child: ListTile(
+                                title: Row(
+                                  mainAxisAlignment: MainAxisAlignment.start,
+                                  children: [
+                                    Text(
+                                      intervals[_resultsList[index]]!
+                                          .elementAt(intervalIndex),
+                                    ),
+                                  ],
+                                ),
+                              ),
+                            );
+                          },
+                        ),
+                      ),
+                    );
                   },
-                  child: const Text("Scan QR")),
+                ),
+              ),
             ],
           ),
-        ));
+        ),
+      ),
+    );
   }
 
   _showScan() async {
     await showDialog(context: context, builder: (context) => const Scan());
-  }
-
-  _showCourses() async {
-    await showDialog(
-        context: context,
-        builder: (context) => AlertDialog(
-              backgroundColor: Colors.lightBlue,
-              title: const Text("Courses\n"),
-              actions: <Widget>[
-                Text(
-                  stringWord,
-                  textAlign: TextAlign.left,
-                )
-              ],
-            ));
   }
 }
